LPS-120601 Move common styles to separated component
